issue #35 delete only empty collections

+++ src/PermissionsModule.php
@@ -211,7 +211,11 @@ class PermissionsModule extends Module {
 
 		if (null !== $deletePermissionCollectionHandler) {
 			foreach (PermissionsCollections::find()->where(['not', 'name' => $currentPermissionsCollectionsNames])->all() as $unusedCollection) {
-				$deletePermissionCollectionHandler($unusedCollection, $unusedCollection->delete());//todo: удаление коллекции удаляет все связи
+				/** @var PermissionsCollections $unusedCollection */
+				if ([] === $unusedCollection->relatedPermissions) {//Нельзя удалять коллекции по имени, нужно удалять те, в которых не осталось правил
+					$deletePermissionCollectionHandler($unusedCollection, $unusedCollection->delete());//todo: удаление коллекции удаляет все связи
+				}
+
 			}
 		} else {
 			PermissionsCollections::deleteAll(['not', 'name' => $currentPermissionsCollectionsNames]);
